Include ./.sapling-templates/ in template search
If there is a directory .sapling-templates present in the current
directory, then that directory will also be searched when searching
for templates.
This allows to define templates that are only relevant for some projects,
without cluttering the list of available templates for other projects.

(Also, I want to see if GitHub recognizes the signature on this commit...)